feat: 优化 tg bot 相关逻辑

+++ server/model/analysis.go
@@ -4,7 +4,7 @@ type OrderSummary struct {
 	Date           string  `json:"date"`
 	OrderTotal     int64   `json:"order_total"`
 	IncomeTotal    float64 `json:"income_total"`
-	GeneralTotal   float64 `json:"general_total"`
+	GeneralTotal   int64   `json:"general_total"`
 	RechargeTotal  int64   `json:"recharge_total"`
 	SubscribeTotal int64   `json:"subscribe_total"`
 }
